add circular imports check
