+++ package.json
@@ -41,6 +41,7 @@
 		"@vitest/ui": "catalog:",
 		"all-contributors-cli": "^6.26.1",
 		"changesets-changelog-clean": "1.3.0",
+		"dpdm": "^4.2.0",
 		"esbuild": "catalog:",
 		"knip": "^6.12.2",
 		"mdxlint": "^1.0.0",
